+++ both/models/Users.js
@@ -1,3 +1,11 @@
+Users = Meteor.users;
+
+if (Meteor.isServer) {
+  Meteor.publish('users', function () {
+    return Users.find();
+  });
+}
+
 User = {
   get: function() {
     return Meteor.user() || {};